+++ jaorm-extensions/jaorm-extension-api/src/main/java/io/github/ulisse1996/jaorm/extension/api/ValidatorExtension.java
@@ -1,0 +1,13 @@
+package io.github.ulisse1996.jaorm.extension.api;
+
+import javax.annotation.processing.ProcessingEnvironment;
+import javax.lang.model.element.Element;
+import java.lang.annotation.Annotation;
+import java.util.Set;
+
+public interface ValidatorExtension {
+
+    Set<Class<? extends Annotation>> getSupported();
+    void validate(Set<Element> elements, ProcessingEnvironment processingEnvironment);
+    void validateSql(String sql, ProcessingEnvironment processingEnvironment);
+}