+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Feature: SL4J support for standard Logging with custom fallback on SPI JaormLoggerHandler
 - Feature: List\<TableRow> support for @Query
 - Feature: Merge DSL
+- Feature: Extension API for Jaorm Processor validation
+- Feature: Default ANSI Sql Query Validator
 
 ## 1.6.0
 - Refactor: Move generation bundle SPI to SPI Provider for Single Feature